move mapBeforeSort into its own file
